requirements.txt
uvicorn
fastapi
pyjwt[crypto]
requests
python-certifi-win32